+++ RS232/mainSerie.c
@@ -1,9 +1,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <unistd.h>
+#include <fcntl.h>
+#include "termset.h"
 
 extern int menuModo();
-extern void modoLocal();
+//extern void modoLocal();
 extern void modoRemoto();
 
 int main (int argc, char *argv[]) {
@@ -15,8 +18,8 @@ int main (int argc, char *argv[]) {
 
     if(mode == 1)
       modoRemoto();
-    else
-      modoLocal();
+    //else
+      //modoLocal();
   }
 
   return 0;
